1. Fix some edit profile bugs 2. Make upload image reusable
