+++ src/components/views/EditPayment.jsx
@@ -11,7 +11,6 @@ import {
   TraceDescription,
   TraceDonateToCommunity,
   TraceFiatAmountCurrency,
-  TracePicture,
   TraceRecipientAddress,
   TraceTitle,
   TraceToken,
@@ -27,6 +26,7 @@ import config from '../../configuration';
 import { Trace } from '../../models';
 import { TraceService } from '../../services';
 import { sendAnalyticsTracking } from '../../lib/SegmentAnalytics';
+import UploadPicture from '../UploadPicture';
 
 const WAIT_INTERVAL = 1000;
 
@@ -445,10 +445,10 @@ function EditPayment(props) {
                     disabled={traceHasFunded}
                   />
 
-                  <TracePicture
+                  <UploadPicture
                     setPicture={setPicture}
-                    traceTitle={payment.title}
                     picture={payment.image}
+                    imgAlt={payment.title}
                   />
 
                   <TraceDonateToCommunity
